Extract Pharmacy and Drug classes to a dedicated folder
Folder named `models` now contains the different classes used across the application, to improve clarity.
