+++ models/Drug.js
@@ -1,0 +1,7 @@
+export default class Drug {
+  constructor(name, expiresIn, benefit) {
+    this.name = name;
+    this.expiresIn = expiresIn;
+    this.benefit = benefit;
+  }
+}
