+++ models/Pharmacy.js
@@ -1,12 +1,4 @@
-export class Drug {
-  constructor(name, expiresIn, benefit) {
-    this.name = name;
-    this.expiresIn = expiresIn;
-    this.benefit = benefit;
-  }
-}
-
-export class Pharmacy {
+export default class Pharmacy {
   constructor(drugs = []) {
     this.drugs = drugs;
   }
