+++ models/index.js
@@ -1,0 +1,2 @@
+export { default as Drug } from "./Drug";
+export { default as Pharmacy } from "./Pharmacy";
