+++ tests/pharmacy.test.js
@@ -1,4 +1,4 @@
-import { Drug, Pharmacy } from "../pharmacy";
+import { Drug, Pharmacy } from "../models";
 
 describe("Pharmacy", () => {
   it("should decrease the benefit and expiresIn", () => {
